+++ study/api-src/org/labkey/api/specimen/importer/EditableSpecimenImporter.java
@@ -13,12 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.labkey.study.importer;
+package org.labkey.api.specimen.importer;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.jetbrains.annotations.NotNull;
-import org.labkey.api.annotations.Migrate;
 import org.labkey.api.collections.CaseInsensitiveHashSet;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.DbSchema;
@@ -30,15 +29,7 @@ import org.labkey.api.security.User;
 import org.labkey.api.specimen.SpecimenColumns;
 import org.labkey.api.specimen.SpecimenRequestManager;
 import org.labkey.api.specimen.SpecimenSchema;
-import org.labkey.api.specimen.importer.EventVialRollup;
-import org.labkey.api.specimen.importer.IteratorSpecimenImportFile;
-import org.labkey.api.specimen.importer.RollupHelper;
 import org.labkey.api.specimen.importer.RollupHelper.RollupMap;
-import org.labkey.api.specimen.importer.RollupInstance;
-import org.labkey.api.specimen.importer.SpecimenColumn;
-import org.labkey.api.specimen.importer.SpecimenImportFile;
-import org.labkey.api.specimen.importer.SpecimenTableType;
-import org.labkey.api.specimen.importer.StandardSpecimenImportStrategy;
 import org.labkey.api.study.SpecimenImportStrategy;
 
 import java.io.IOException;
@@ -53,7 +44,6 @@ import java.util.Set;
  * Date: 6/13/13
  * Time: 12:13 PM
  */
-@Migrate // Depends only on SpecimenImporter
 public class EditableSpecimenImporter extends SpecimenImporter
 {
     private static final String GUID_COLNAME = "GlobalUniqueId";
