+++ study/api-src/org/labkey/api/study/model/VisitService.java
@@ -1,5 +1,8 @@
 package org.labkey.api.study.model;
 
+import org.apache.logging.log4j.Logger;
+import org.jetbrains.annotations.Nullable;
+import org.labkey.api.security.User;
 import org.labkey.api.services.ServiceRegistry;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.Visit;
@@ -23,4 +26,11 @@ public interface VisitService
     }
 
     List<? extends Visit> getVisits(Study study, Visit.Order order);
+
+    void updateParticipantVisitsWithCohortUpdate(Study study, User user, @Nullable Logger logger);
+
+    /**
+     * Updates this study's participant, visit, and participant visit tables. Also updates automatic cohort assignments.
+     */
+    void updateParticipantVisits(Study study, User user);
 }
