+++ study/src/org/labkey/study/controllers/StudyDesignController.java
@@ -37,6 +37,7 @@ import org.labkey.api.security.permissions.UpdatePermission;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.TimepointType;
 import org.labkey.api.study.Visit;
+import org.labkey.api.study.security.permissions.ManageStudyPermission;
 import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.HttpView;
 import org.labkey.api.view.JspView;
@@ -58,7 +59,6 @@ import org.labkey.study.model.TreatmentManager;
 import org.labkey.study.model.TreatmentProductImpl;
 import org.labkey.study.model.TreatmentVisitMapImpl;
 import org.labkey.study.model.VisitImpl;
-import org.labkey.api.study.security.permissions.ManageStudyPermission;
 import org.labkey.study.visitmanager.VisitManager;
 import org.springframework.validation.BindException;
 import org.springframework.validation.Errors;
