+++ study/src/org/labkey/study/controllers/designer/DesignerController.java
@@ -48,6 +48,7 @@ import org.labkey.api.security.permissions.AdminPermission;
 import org.labkey.api.security.permissions.DeletePermission;
 import org.labkey.api.security.permissions.ReadPermission;
 import org.labkey.api.security.permissions.UpdatePermission;
+import org.labkey.api.specimen.importer.SimpleSpecimenImporter;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
 import org.labkey.api.study.TimepointType;
@@ -71,7 +72,6 @@ import org.labkey.study.designer.StudyDesignManager;
 import org.labkey.study.designer.StudyDesignVersion;
 import org.labkey.study.designer.XMLSerializer;
 import org.labkey.study.designer.view.StudyDesignsWebPart;
-import org.labkey.study.importer.SimpleSpecimenImporter;
 import org.labkey.study.view.StudyGWTView;
 import org.labkey.study.view.VaccineStudyWebPart;
 import org.springframework.validation.BindException;
