+++ study/src/org/labkey/study/controllers/specimen/SpecimenController.java
@@ -95,6 +95,7 @@ import org.labkey.api.specimen.actions.ManageReqsBean;
 import org.labkey.api.specimen.actions.ReportConfigurationBean;
 import org.labkey.api.specimen.actions.SpecimenReportActions;
 import org.labkey.api.specimen.importer.RequestabilityManager;
+import org.labkey.api.specimen.importer.SimpleSpecimenImporter;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
 import org.labkey.api.specimen.model.ExtendedSpecimenRequestView;
@@ -176,7 +177,6 @@ import org.labkey.study.StudySchema;
 import org.labkey.study.controllers.BaseStudyController;
 import org.labkey.study.controllers.DatasetController;
 import org.labkey.study.designer.MapArrayExcelWriter;
-import org.labkey.study.importer.SimpleSpecimenImporter;
 import org.labkey.study.model.DatasetDefinition;
 import org.labkey.study.model.SecurityType;
 import org.labkey.study.model.StudyImpl;
